requirements.txt
pyannote.audio
pydub
librosa
soundfile 